requirements.txt
Flask
psycopg2-binary
pandas
scikit-learn
joblib
SQLAlchemy
